[HACK] efi: libstub: avoid ilog2(u64)
drivers/firmware/efi/libstub/random.stub.o: In function `efi_random_alloc':
random.c:(.text.efi_random_alloc+0x1c0): undefined reference to `__aeabi_llsr'

This change is safe because the alignment is always a small number

+++ drivers/firmware/efi/libstub/random.c
@@ -101,7 +101,7 @@ efi_status_t efi_random_alloc(efi_system_table_t *sys_table_arg,
 		efi_memory_desc_t *md = (void *)memory_map + map_offset;
 		unsigned long slots;
 
-		slots = get_entry_num_slots(md, size, ilog2(align));
+		slots = get_entry_num_slots(md, size, ilog2((u32)align));
 		MD_NUM_SLOTS(md) = slots;
 		total_slots += slots;
 	}
